// frontend/src/api-client/types.gen.ts
export type BackendInternalModelForm = {
    createdAt?: string;
    description?: string;
    fields?: Array<BackendInternalModelFormField>;
    id?: string;
    published?: boolean;
    responses?: Array<BackendInternalModelFormResponse>;
    slug?: string;
    title?: string;
    updatedAt?: string;
    userId?: string;
};

export type BackendInternalModelFormField = {
    createdAt?: string;
    fieldType?: string;
    formId?: string;
    id?: string;
    label?: string;
    /**
     * @swaggertype object
     */
    options?: Array<number>;
    orderIndex?: number;
    required?: boolean;
    updatedAt?: string;
};

export type BackendInternalModelFormResponse = {
    /**
     * @swaggertype object
     */
    answers?: Array<number>;
    formId?: string;
    id?: string;
    submittedAt?: string;
};